origin/entidad-animalXusuario

+++ src/main/java/com/grupo3/pawHome/repositories/UsuarioRepository.java
@@ -1,7 +1,10 @@
 package com.grupo3.pawHome.repositories;
 
 import com.grupo3.pawHome.entities.Usuario;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+import java.util.Optional;
+
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 }
